factoring out chain logic
