+++ chain/beacon/constants.go
@@ -1,4 +1,4 @@
-package chain
+package beacon
 
 import "time"
