+++ go.mod
@@ -4,6 +4,7 @@ go 1.14
 
 require (
 	github.com/BurntSushi/toml v0.3.1
+	github.com/boltdb/bolt v1.3.1
 	github.com/drand/bls12-381 v0.3.2
 	github.com/drand/kyber v1.0.2
 	github.com/go-kit/kit v0.10.0
